js: git rid of ts-loader by moving tests in src directory
