+++ js/package.json
@@ -32,10 +32,6 @@
     "@types/mocha": "^2.2.48",
     "@types/requirejs": "^2.1.31",
     "@types/three": "^0.92.4",
-    "babel-loader": "7.1.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-es2015-without-strict": "0.0.4",
-    "cache-loader": "^1.2.2",
     "chai": "^4.1.2",
     "eslint": "^5.13.0",
     "eslint-config-standard": "^12.0.0",
@@ -51,8 +47,6 @@
     "karma-mocha-reporter": "^2.2.5",
     "karma-sinon": "^1.0.5",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-typescript": "^3.0.12",
-    "karma-typescript-es6-transform": "^1.0.4",
     "karma-webpack": "^3.0.5",
     "mocha": "^3.5.3",
     "npm-run-all": "^4.1.3",
@@ -60,8 +54,6 @@
     "sinon": "^2.4.1",
     "sinon-chai": "^2.11.0",
     "source-map-loader": "^0.2.4",
-    "thread-loader": "^1.2.0",
-    "ts-loader": "^4.4.2",
     "typescript": "~3.0.0",
     "url-loader": "^0.5.7",
     "webpack": "^4.29.1",
@@ -89,7 +81,8 @@
     "dist/",
     "css/",
     "src/",
-    "lib/",
+    "lib/*.js",
+    "lib/three/*.js",
     "data/",
     "glsl/"
   ]
