+++ js/src/test/dummy-manager.ts
