+++ js/src/test/figure.js
@@ -1,4 +1,4 @@
-import { FigureModel, FigureView, WidgetManagerHackModel } from "..";
+import { FigureModel, FigureView, WidgetManagerHackModel } from "../..";
 import { expect } from 'chai';
 
 describe("figure >", () => {
