+++ js/tsconfig.json
@@ -9,6 +9,7 @@
     "target": "es2015",
     "outDir": "lib",
     "skipLibCheck": true,
+    "strictNullChecks": false,
     // "sourceMap": true,
     // inline seems to work better with karma
     "inlineSourceMap": true,
